refactor: Consolidate improved implementations as defaults
- Replace algorithms.py with improved PPO implementation
- Remove environment_improved.py (functionality integrated)
- Rename ImprovedPPOAgent to PPOAgent as default
- Add backward compatibility types for existing tests
- Update documentation references

This consolidation eliminates the "_improved" naming pattern
and makes the enhanced implementations the standard defaults.

Breaking changes:
- PPOConfig structure updated with new fields
- Some test interfaces may need updates
